clippy: make --all-targets green under -D warnings (test modules + src)
The prior green bar ran plain `cargo clippy` (lib+bins only); enabling
benches required --all-targets, which also lints #[cfg(test)] code and
surfaced 97 pre-existing lints. Resolved all of them:

- cargo clippy --fix for the machine-applicable bulk (collapsible_if →
  let-chains, redundant_field_names, unnecessary_cast, etc.)
- by hand: manual_checked_ops → checked_div (metrics, provider_health),
  sort_by → sort_by_key+Reverse (analytics), get().is_some() →
  contains_key (rollback), nested format! flatten (search),
  type alias for complex plugin-registry type, field_reassign_with_default
  → struct-update syntax (test_hub), drop tautological assert!(true).

Behavior-preserving throughout; cargo test --workspace still 652 passed,
0 failed. clippy/fmt/check all clean under --all-targets.

Note: provider_health::summary() mapped both `unhealthy>0` and
`degraded>0` to Degraded (collapsed, behavior kept) — worth confirming
whether unhealthy providers should yield Unhealthy overall.

+++ prompt-hub/src/budget.rs
@@ -130,7 +130,7 @@ impl BudgetTracker {
         self.set_budget(config.monthly_budget_usd);
         let threshold_micros = (config.alert_threshold_percent * 1_000_000.0) as u64;
         self.alert_threshold_percent
-            .store((threshold_micros / 1_000_000) as u64, Ordering::SeqCst);
+            .store(threshold_micros / 1_000_000, Ordering::SeqCst);
         Ok(())
     }
